Move constants.js
