+++ src/index.js
@@ -1,8 +1,8 @@
 import path from 'path';
 
-import {ERROR_MODULE_NOT_FOUND} from './constants';
 import {doTransform} from './do-transform';
 import {getLoaderOptions} from './get-loader-options';
+import {ERROR_MODULE_NOT_FOUND} from './lib/constants';
 import {getImportPath} from './utils/get-import-path';
